# cmd/api/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/admin/orders": {
            "get": {
                "description": "admin can see all orders in application",
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to show all order",
                "operationId": "GetAllShopOrders",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully order list got",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get shop order data",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/admin/orders/": {
            "put": {
                "description": "admin can change User Orders status",
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to change the status of order",
                "operationId": "UdateOrderStatus",
                "parameters": [
                    {
                        "description": "input field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqUpdateOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully got order items",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/admin/orders/returns": {
            "get": {
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to see all order reutns",
                "operationId": "GetAllOrderReturns",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully got all order returns",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get order returns",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/admin/orders/returns/pending": {
            "get": {
                "description": "admin can see the pending return request and accept it or not",
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to show pending return request and update it",
                "operationId": "GetAllPendingReturns",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully got  pending orders return request",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get pending order return requests",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "put": {
                "description": "admin can approve, cancell etc. updation on User Orders_return",
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to supdate the order_return request from user",
                "operationId": "UpdategReturnRequest",
                "parameters": [
                    {
                        "description": "Input Fiields",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqUpdatReturnOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully order_response updated",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
        "/admin/orders/statuses": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "Admin Orders"
                ],
                "summary": "api for admin to see all order statues for changing order's statuses",
                "operationId": "GetAllOrderStatuses",
                "responses": {
                    "200": {
                        "description": "successfully got all order statueses",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get order statuses",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/carts": {
            "get": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can see all productItem that stored in cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for get all cart item of user",
                "operationId": "User Cart",
                "responses": {
                    "200": {
                        "description": "successfully got user cart items",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get cart items",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can inrement or drement count of a productItem in cart (min=1)",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for updte productItem count",
                "operationId": "UpdateCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCartCount"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem count change on cart"
                    },
                    "400": {
                        "description": "invalid input"
                    },
                    "500": {
                        "description": "can't update the count of product item on cart"
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can add a stock in product to user cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for add productItem to user cart",
                "operationId": "AddToCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem added to cart"
                    },
                    "400": {
                        "description": "can't add the product item into cart"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "description": "user can remove a signle productItem full quantity from cart",
                "tags": [
                    "User Cart"
                ],
                "summary": "api for remove a product from cart",
                "operationId": "RemoveFromCart",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqCart"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successfully productItem removed from cart",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "can't remove product item from cart",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/carts/place-order/cod": {
            "post": {
                "security": [
                    {
                        "ApiKeyAuth": []
                    }
                ],
                "tags": [
                    "User Cart"
                ],
                "summary": "api for user to place an order on cart with COD",
                "operationId": "PlaceOrderCartCOD",
                "parameters": [
                    {
                        "description": "Input Field",
                        "name": "inputs",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/req.ReqPlaceOrder"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully order placed in COD",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to save shop order",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
        "/orders": {
            "get": {
                "description": "user can see all User Orders history",
                "tags": [
                    "User Orders"
                ],
                "summary": "api for showing User Orders list",
                "operationId": "GetUserOrder",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Page Number",
                        "name": "page_number",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "Count Of Order",
                        "name": "count",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successfully got shop order list of user",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "500": {
                        "description": "faild to get user shop order list",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            },
            "post": {
                "description": "user can cancell the order if it's not placed",
                "tags": [
                    "User Orders"
                ],
                "summary": "api for user to cancell the order",
                "operationId": "CancellOrder",
                "responses": {
                    "200": {
                        "description": "Successfully order cancelled",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    },
                    "400": {
                        "description": "invalid input on param",
                        "schema": {
                            "$ref": "#/definitions/res.Response"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "req.ReqCart": {
            "type": "object",
            "required": [
                "product_item_id"
            ],
            "properties": {
                "product_item_id": {
                    "type": "integer"
                }
            }
        },
        "req.ReqCartCount": {
            "type": "object",
            "required": [
                "product_item_id"
            ],
            "properties": {
                "count": {
                    "type": "integer",
                    "minimum": 1
                },
                "product_item_id": {
                    "type": "integer"
                }
            }
        },
...
        "req.ReqPlaceOrder": {
            "type": "object",
            "required": [
                "address_id",
                "payment_method_id"
            ],
            "properties": {
                "address_id": {
                    "type": "integer"
                },
                "payment_method_id": {
                    "type": "integer"
                }
            }
        },
...
        "req.ReqUpdatReturnOrder": {
            "type": "object",
            "required": [
                "order_return_id",
                "order_status_id"
            ],
            "properties": {
                "admin_comment": {
                    "type": "string"
                },
                "order_return_id": {
                    "type": "integer"
                },
                "order_status_id": {
                    "type": "integer"
                }
            }
        },
        "req.ReqUpdateOrder": {
            "type": "object",
            "required": [
                "shop_order_id"
            ],
            "properties": {
                "order_status_id": {
                    "type": "integer"
                },
                "shop_order_id": {
                    "type": "integer"
                }
            }
        },
...
        "res.Response": {
            "type": "object",
            "properties": {
                "data": {},
                "errors": {},
                "message": {
                    "type": "string"
                },
                "status_code": {
                    "type": "integer"
                }
            }
        }
    }
}